moved and reorganized files and some definitions

close to eliminating warning problems and preparing compilers for future package management efforts


non working docker files


update a couple of things


forgot one thing with the cache


switch all logs to eris cli logs
