+++ glide.yaml
@@ -1,9 +1,10 @@
 package: github.com/eris-ltd/eris-compilers
 import:
-- package: github.com/ebuchman/go-shell-pipes
 - package: github.com/eris-ltd/common
   subpackages:
-  - go/common
   - go/docs
+- package: github.com/eris-ltd/eris-cli
+  subpackages:
+  - config
 - package: github.com/eris-ltd/eris-logger
 - package: github.com/spf13/cobra
